This is 254-8 as posted on facepunch forums:
I added an 'Import Textures' button to the Material property page. Import a model, select the mesh object, then scroll to the bottom of the Material page and click the button in the 'SMD Import' panel. That lets you choose a directory to load images from.

It will load an image for each material according to the material's name - or 'smd_name' in the Custom Properites panel if the original material name was longer than 21 characters (a Blender limit). If that name doesn't have a file extension or the file doesn't exist then it will try 'name.tga' followed by 'name.bmp'.

There's no harm clicking the button multiple times, the image file won't be loaded if it already was, and won't be assigned to a material texture slot if it already was.