Fix issues with importing animations when the animation has a different hierarchy than the target armature.
Ex 1: sourcesdk_content\hl2\modelsrc\weapons\v_smg1\Alt_fire.smd
the animation has an extra root bone.

Ex 2: sourcesdk_content\hl2\modelsrc\Buggy\buggy_ammo_open.smd
the mesh has an extra bone Gun_Parent.

When the pose armature has extra ancestors, those ancestors
are ignored when setting the pose.  Instead a common ancestor
is found, if any, and the pose is set relative to that common
ancestor (as in Ex 1 above).

When the target armature has extra ancestors, those ancestors
are added to the pose armature (as in Ex 2 above) and the pose
is set relative to the first of those ancestors rather than the
ancestor in the pose armature.

-----

Improved the test suite to compare the global pos/rot of bones
rather than just the parent-relative pos/rot stored in the smd.

Give hints about bones we know are added/removed when running a test
to avoid printing out a bunch of spurious error messages.